+++ RELEASE_NOTES.md
@@ -3,7 +3,7 @@
 * [#8](https://github.com/skuzzle/snapshot-tests/issues/8): Way more sophisticated orphan detection
 * Deprecate `EnableSnapshotTests.forceUpdateSnapshots` in favor of new `@ForceUpdateSnapshots` annotation
 * Orphaned files are not deleted any more when `forceUpdateSnapshots` is enabled
-* `@DeleteOrphans` annotation can be placed on the test class in order to delete orphaned files
+* `@DeleteOrphanedSnapshots` annotation can be placed on the test class in order to delete orphaned files
 
 Maven Central coordinates for this release: